+++ osrframework/enumeration.py
@@ -3,7 +3,7 @@
 #
 ##################################################################################
 #
-#    Copyright 2015 Félix Brezo and Yaiza Rubio (i3visio, contacto@i3visio.com)
+#    Copyright 2016 Félix Brezo and Yaiza Rubio (i3visio, contacto@i3visio.com)
 #
 #    This program is part of OSRFramework. You can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -21,7 +21,7 @@
 ##################################################################################
 
 '''
-enumeration.py Copyright (C) F. Brezo and Y. Rubio (i3visio) 2015
+enumeration.py Copyright (C) F. Brezo and Y. Rubio (i3visio) 2016
 This program comes with ABSOLUTELY NO WARRANTY.
 This is free software, and you are welcome to redistribute it under certain conditions. For additional info, visit to <http://www.gnu.org/licenses/gpl-3.0.txt>.
 '''
@@ -33,7 +33,7 @@ import osrframework.utils.browser as browser
 import re
 
 __author__ = "Felix Brezo, Yaiza Rubio "
-__copyright__ = "Copyright 2015, i3visio"
+__copyright__ = "Copyright 2016, i3visio"
 __credits__ = ["Felix Brezo", "Yaiza Rubio"]
 __license__ = "GPLv3+"
 __version__ = "v1.0.2"
@@ -85,7 +85,7 @@ def enumeration_main(args):
     '''
         Main loop.
     '''
-    sayingHello = """enumeration.py Copyright (C) F. Brezo and Y. Rubio (i3visio) 2015
+    sayingHello = """enumeration.py Copyright (C) F. Brezo and Y. Rubio (i3visio) 2016
 This program comes with ABSOLUTELY NO WARRANTY.
 This is free software, and you are welcome to redistribute it under certain conditions. For additional info, visit <http://www.gnu.org/licenses/gpl-3.0.txt>."""
     #logger.info(sayingHello)